added gizmos

+++ GameEngine/src/Engine/Renderer/Renderer2D.cpp
@@ -141,7 +141,7 @@ namespace Engine
 	{
 		CREATE_PROFILE_FUNCTIONI();
 
-		uint32_t dataSize = (uint8_t*)s_Data.QuadVertexBufferPtr - (uint8_t*)s_Data.QuadVertexBufferBase;
+		uint32_t dataSize = (uint32_t)((uint8_t*)s_Data.QuadVertexBufferPtr - (uint8_t*)s_Data.QuadVertexBufferBase);
 		s_Data.QuadVertexBuffer->SetData(s_Data.QuadVertexBufferBase, dataSize);
 
 		Flush();
@@ -151,7 +151,7 @@ namespace Engine
 	{
 		CREATE_PROFILE_FUNCTIONI();
 
-		for (int i = 0; i < s_Data.TextureSlotIndex; i++)
+		for (uint32_t i = 0; i < s_Data.TextureSlotIndex; i++)
 		{
 			s_Data.TextureSlots[i].get()->Bind(i);
 		}
@@ -346,7 +346,7 @@ namespace Engine
 
 		float textureIndex = 0.0f;
 
-		for (int i = 1; i < s_Data.TextureSlotIndex; i++)
+		for (uint32_t i = 1; i < s_Data.TextureSlotIndex; i++)
 		{
 			if (s_Data.TextureSlots[i].get() == texture.get())
 			{
